refactor: moves PropertyRegister page
